% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_documentation.R
\docType{package}
\name{ccoptimalmatch}
\alias{ccoptimalmatch}
\alias{ccoptimalmatch-package}
\title{ccoptimalmatch: Optimal Case Control matching}
\description{
Fast and optimal matching for cases and controls
}
\author{
\strong{Maintainer}: Pavlos Mamouris \email{pavlos.mamouris@kuleuven.be}

Authors:
\itemize{
  \item Vahid Nassiri \email{vahid.nassiri@openanalytics.eu} [contributor]
}

}
