% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interations_per_round_dt.R
\name{iterations_per_round_dt}
\alias{iterations_per_round_dt}
\title{Iterations per Round (data.table)}
\usage{
iterations_per_round_dt(
  dataset,
  cluster_var,
  Id_Patient,
  total_cont_per_case,
  case_control,
  mat_per_case = NULL
)
}
\arguments{
\item{dataset}{a dataset}

\item{cluster_var}{a string}

\item{Id_Patient}{a string}

\item{total_cont_per_case}{a string}

\item{case_control}{a string}

\item{mat_per_case}{a numeric indicating number of matches per case with default of \code{NULL}}
}
\description{
Iterations per Round (data.table)
}
\keyword{internal}
